feat: adds template command
