+++ .github/workflows/deploy.yml
@@ -67,7 +67,7 @@ jobs:
       - name: Deploy
         uses: input-output-hk/catalyst-forge/actions/run@master
         with:
-          command: deploy
+          command: deploy push
           args: ${{ matrix.deployment }}
           local: ${{ inputs.local }}
           verbosity: ${{ inputs.verbosity }}
