+++ cli/cmd/cmds/deploy/cmd.go
@@ -1,0 +1,6 @@
+package deploy
+
+type DeployCmd struct {
+	Push     PushCmd     `cmd:"" help:"Pushes a project deployment to the GitOps repo."`
+	Template TemplateCmd `cmd:"" help:"Generates a project's deployment YAML."`
+}
