+++ cli/cmd/cmds/deploy/deploy.go
@@ -1,4 +1,4 @@
-package cmds
+package deploy
 
 import (
 	"fmt"
@@ -7,11 +7,11 @@ import (
 	"github.com/input-output-hk/catalyst-forge/cli/pkg/run"
 )
 
-type DeployCmd struct {
+type PushCmd struct {
 	Project string `arg:"" help:"The path to the project to deploy." kong:"arg,predictor=path"`
 }
 
-func (c *DeployCmd) Run(ctx run.RunContext) error {
+func (c *PushCmd) Run(ctx run.RunContext) error {
 	project, err := ctx.ProjectLoader.Load(c.Project)
 	if err != nil {
 		return fmt.Errorf("could not load project: %w", err)
