+++ cli/cmd/cmds/secret.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 
 	"github.com/input-output-hk/catalyst-forge/cli/pkg/run"
+	"github.com/input-output-hk/catalyst-forge/cli/pkg/utils"
 	"github.com/input-output-hk/catalyst-forge/lib/project/schema"
 	"github.com/input-output-hk/catalyst-forge/lib/project/secrets"
 )
@@ -99,7 +100,7 @@ func (c *Get) Run(ctx run.RunContext) error {
 			fmt.Println(mappedSecret[c.Key])
 			return nil
 		} else {
-			printJson(mappedSecret, false)
+			utils.PrintJson(mappedSecret, false)
 			return nil
 		}
 	}
